+++ Mods ChoGGi/Map Overview More Info/metadata.lua
@@ -1,13 +1,12 @@
 return PlaceObj("ModDef", {
---~ 	"title", "Map Overview Show Surface Resources",
 	"title", "Map Overview More Info",
 	"id", "ChoGGi_MapOverviewShowSurfaceResources",
-	"lua_revision", 1001569,
 	"steam_id", "1768449416",
 	"pops_any_uuid", "77783c38-6f89-4371-9628-a6fdb2fec8bb",
-	"version", 6,
+	"lua_revision", 1001569,
+	"version", 7,
 	"version_major", 0,
-	"version_minor", 6,
+	"version_minor", 7,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {
@@ -20,6 +19,7 @@ Scan progress as overlay text to map sectors.
 
 Mod Options:
 Show Metals/Show Polymers/Show Scan Progress: What you'd expect.
+Show Dropped Resources: Show indicator for any dropped resource piles (this will ignore any within range of a drone controller).
 Text Opacity: 0-255 (0 == completely visible).
 Text Background: Add black background around info.
 Text Style: Defaults to a larger text style, change with this (see tooltip for text styles).