+++ Mods ChoGGi/Persistent Transport Route/metadata.lua
@@ -1,21 +1,21 @@
-return PlaceObj("ModDef", {
-	"title", "Persistent Transport Route",
-	"version", 1,
-	"version_major", 0,
-	"version_minor", 1,
-	"image", "Preview.png",
-	"id", "ChoGGi_PersistentTransportRoute",
-	"steam_id", "1804585985",
-	"pops_any_uuid", "92513d0f-8b09-4f82-82e9-246511849a04",
-	"author", "ChoGGi",
-	"lua_revision", 1001569,
-	"code", {
-		"Code/Script.lua",
-	},
-	"TagGameplay", true,
-	"description", [[Zooming to the map overview will no longer cancel your transport route.
-If you haven't set a from location, then it'll just reset like usual because I don't care.
+--~ return PlaceObj("ModDef", {
+--~ 	"title", "Persistent Transport Route",
+--~ 	"id", "ChoGGi_PersistentTransportRoute",
+--~ 	"steam_id", "1804585985",
+--~ 	"pops_any_uuid", "92513d0f-8b09-4f82-82e9-246511849a04",
+--~ 	"lua_revision", 1001569,
+--~ 	"version", 1,
+--~ 	"version_major", 0,
+--~ 	"version_minor", 1,
+--~ 	"image", "Preview.png",
+--~ 	"author", "ChoGGi",
+--~ 	"code", {
+--~ 		"Code/Script.lua",
+--~ 	},
+--~ 	"TagGameplay", true,
+--~ 	"description", [[Zooming to the map overview will no longer cancel your transport route.
+--~ If you haven't set a from location, then it'll just reset like usual because I don't care.
 
 
-Requested by: Pretty much anyone that's ever played the game (eh devs...).]],
-})
+--~ Requested by: Pretty much anyone that's ever played the game (eh devs...).]],
+--~ })