+++ Mods ChoGGi/Renegade Progress/metadata.lua
@@ -1,15 +1,14 @@
 return PlaceObj("ModDef", {
 	"title", "Renegade Progress",
-	"version", 3,
-	"version_major", 0,
-	"version_minor", 3,
-
-	"image", "Preview.png",
 	"id", "ChoGGi_RenegadeProgress",
 	"steam_id", "1802131001",
 	"pops_any_uuid", "ce2f0465-bec8-48d1-acab-9656b9c5907e",
-	"author", "ChoGGi",
 	"lua_revision", 1001569,
+	"version", 3,
+	"version_major", 0,
+	"version_minor", 3,
+	"image", "Preview.png",
+	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},