+++ Mods ChoGGi/Tourists Remove Specs/Code/Script.lua
@@ -1,16 +1,18 @@
 -- See LICENSE for terms
 
 local specs
+local function UpdateSpecs()
+	specs = table.copy(const.ColonistSpecialization)
+	specs.Tourist = nil
+	specs.none = nil
+end
 
 local mod_EnableMod
 
 -- fired when settings are changed/init
 local function ModOptions()
 	mod_EnableMod = CurrentModOptions:GetProperty("EnableMod")
-
-	specs = table.copy(const.ColonistSpecialization)
-	specs.Tourist = nil
-	specs.none = nil
+	UpdateSpecs()
 end
 
 -- load default/saved settings
@@ -32,10 +34,12 @@ local function UpdateColonist(obj)
 				-- tourist with a spec
 				obj.city:RemoveFromLabel(trait, obj)
 				obj:RemoveTrait(trait)
-				-- I guess I could, I mean devs add it to ColonistSpecialization
+				-- I guess I could, I mean devs added it to ColonistSpecialization
 				obj.specialist = "Tourist"
 --~ 				obj.specialist = "none"
 				obj:ChooseEntity()
+			elseif trait == "Tourist" then
+				obj.specialist = "Tourist"
 			end
 		end
 	end
@@ -43,10 +47,11 @@ end
 
 GlobalVar("g_ChoGGi_TouristsRemoveSpecs_StartupClean", false)
 local function StartupCode()
-	if not mod_EnableMod or g_ChoGGi_TouristsRemoveSpecs_StartupClean then
+	if not mod_EnableMod or g_ChoGGi_TouristsRemoveSpecs_StartupClean or rawget(_G, "g_AT_Options") then
 		return
 	end
 
+	UpdateSpecs()
 	local objs = UICity.labels.Colonist or ""
 	for i = 1,# objs do
 		UpdateColonist(objs[i])
@@ -60,7 +65,8 @@ OnMsg.LoadGame = StartupCode
 
 -- update newly arrived
 function OnMsg.ColonistArrived(obj)
-	if mod_EnableMod then
+	if mod_EnableMod or not rawget(_G, "g_AT_Options") then
+		UpdateSpecs()
 		UpdateColonist(obj)
 	end
 end