+++ Mods ChoGGi/Tourists Remove Specs/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2431912606",
 	"pops_any_uuid", "5cabd295-b917-4324-ac87-f84a54355dfc",
 	"lua_revision", 1001569,
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
@@ -17,6 +17,7 @@ return PlaceObj("ModDef", {
 	"description", [[Remove specializations from all tourists, so filters work a little better.
 Mod option to disable (only applies to new tourists, existing have specs removed).
 
+This mod does nothing if Automated Tourism is installed.
 
 Requested by pikasnoop.
 ]],