+++ Mods ChoGGi/Triboelectric Scrubbers Clean Dome Buildings/Code/Script.lua
@@ -5,17 +5,17 @@ function TriboelectricScrubber:CleanBuildings()
 	end
 
 	self:ForEachDirtyInRange(function(dirty, self)
-		if dirty:IsKindOf("Building") then
-			if dirty ~= self then
-				if dirty:IsKindOf("DustGridElement") then
-					dirty:AddDust(-self.dust_clean)
---~ 				elseif not dirty.parent_dome then --outside of dome
-				else
-					dirty:AccumulateMaintenancePoints(-self.dust_clean)
-				end
+		if dirty ~= self then
+			if dirty:IsKindOf("Building") then
+					if dirty:IsKindOf("DustGridElement") then
+						dirty:AddDust(-self.dust_clean)
+--~ 					elseif not dirty.parent_dome and dirty.accumulate_maintenance_points then
+					elseif dirty.accumulate_maintenance_points then
+						dirty:AccumulateMaintenancePoints(-self.dust_clean)
+					end
+			elseif dirty:IsKindOf("DroneBase") then
+				dirty:AddDust(-self.dust_clean)
 			end
-		elseif dirty:IsKindOf("DroneBase") then
-			dirty:AddDust(-self.dust_clean)
 		end
 	end, self)
 end